Rename abortion.ts.
