+++ package/src/index.ts
@@ -4,7 +4,7 @@ export {
     createAbortError,
     isAbortError,
     throwIfAbortError,
-} from './abortion'
+} from './abort'
 export { impreciseBadge } from './badges'
 export { Handler, HandlerArgs } from './handler'
 export {
